requirements.txt
Django==1.9.7
ipdb==0.10.1
lxml==3.6.0
react-render-client==1.0.0
requests==2.10.0
asgi-redis==0.13.1
channels==0.16.0
drf-url-filters==0.1.3